go/cousins-in-binary-tree_test.go
package leetcode

import "testing"

func TestFindDepth(t *testing.T) {
	if isCousins(&TreeNode{
		Val: 1,
		Left: &TreeNode{
			Val: 2,
			Left: &TreeNode{
				Val: 4,
			},
		},
		Right: &TreeNode{
			Val: 3,
		},
	}, 4, 3) != false {
		t.Fatal()
	}
	if isCousins(&TreeNode{
		Val: 1,
		Left: &TreeNode{
			Val: 2,
			Right: &TreeNode{
				Val: 4,
			},
		},
		Right: &TreeNode{
			Val: 3,
			Right: &TreeNode{
				Val: 5,
			},
		},
	}, 5, 4) != true {
		t.Fatal()
	}
	if isCousins(&TreeNode{
		Val: 1,
		Left: &TreeNode{
			Val: 2,
			Right: &TreeNode{
				Val: 4,
			},
		},
		Right: &TreeNode{
			Val: 3,
		},
	}, 2, 3) != false {
		t.Fatal()
	}
}
